_FUM_Training/requirements.txt
torch
norse
tqdm
pyyaml
pynvml
opencv-python
numpy
librosa
